Move DefaultMessages to localization package
