+++ src/main/java/com/freya02/botcommands/api/core/DefaultMessagesSupplier.java
@@ -1,9 +1,9 @@
 package com.freya02.botcommands.api.core;
 
-import com.freya02.botcommands.api.DefaultMessages;
 import com.freya02.botcommands.api.core.config.BServiceConfigBuilder;
 import com.freya02.botcommands.api.core.service.annotations.BService;
 import com.freya02.botcommands.api.core.service.annotations.InterfacedService;
+import com.freya02.botcommands.api.localization.DefaultMessages;
 import net.dv8tion.jda.api.interactions.DiscordLocale;
 import org.jetbrains.annotations.NotNull;
